Rename MyHomePage to LindenScreen and move to the separate file
